+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### 3.2 ###
 * Added: [GFM](https://help.github.com/articles/creating-and-highlighting-code-blocks/) style Markdown support for fenced code blocks
+* Added: Markdown fenced code blocks bbPress support
 * Added: Environment Check to ensure Enlighter is working in a well configured environment
 * Added: Filter hook `enlighter_startup` to disable Enlighter on selected pages - feature requested on [WordPress.org Forums](https://wordpress.org/support/topic/best-way-to-dequeue-enlighter-plugin?replies=2) #43
 * Added: Filter `enlighter_inline_javascript` - applied to inline javascript which is injected into the page